wlcore: dt: parse some additional properties (INTERNAL)
the optional clock properties are used only for wl12xx

convert platform_quirks from unsigned long to u32,
so of_property_read_u32 could be called directly.

+++ include/linux/wl12xx.h
@@ -59,7 +59,7 @@ struct wl12xx_platform_data {
 	int irq;
 	int board_ref_clock;
 	int board_tcxo_clock;
-	unsigned long platform_quirks;
+	u32 platform_quirks;
 	bool pwr_in_suspend;
 };
